Refatorar método para exibir todos os funcionários ativos na Gestor Controller.
