+++ src/main/java/br/com/zup/PointMarker/gestor/Gestor.java
@@ -1,19 +1,11 @@
 package br.com.zup.PointMarker.gestor;
 
-import br.com.zup.PointMarker.cargo.Cargo;
-import br.com.zup.PointMarker.enums.Status;
-
-import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 import javax.validation.constraints.NotBlank;
-import java.time.LocalDate;
 
 @Entity
 @Table(name = "gestor")
